Added assertion of text boxes in Agents Menu

+++ Lab4/ZadDomLab4JakubKazimierski/ZadDomLab4JakubKazimierski/AgentsMenu.cs
@@ -15,13 +15,13 @@ namespace ZadDomLab4JakubKazimierski
     public partial class AgentsMenu : Form
     {
         private readonly IFBIgeneric<FBIagentsList> _agents;
-        private readonly IFBIgeneric<FBIagentsType> _agentsType;
+      
         public AgentsMenu()
         {
             InitializeComponent();
 
             _agents = new FBIgeneric<FBIagentsList>();
-            _agentsType = new FBIgeneric<FBIagentsType>();
+            
 
             LoadAgents();
         }
@@ -33,23 +33,30 @@ namespace ZadDomLab4JakubKazimierski
 
         private void InsertButton_Click(object sender, EventArgs e)
         {
-            var agentFirstName = InsertNameTextButton.Text;
-            var agentLastName = InsertSurnameTextBox.Text;
-            var agentBirthdate = InsertBirthTextBox.Text;
-            var TypeId = InsertRankTextBox.Text;
-
-            FBIagentsList newAgent = new FBIagentsList()
+            if (InsertNameTextButton.Text != "" && InsertSurnameTextBox.Text != "" && InsertRankTextBox.Text != "" && InsertBirthTextBox.Text != "")
             {
-                FirstName = agentFirstName,
-                LastName = agentLastName,
-                Birthdate = Convert.ToDateTime(agentBirthdate),
-                AgentTypeId = Int32.Parse(TypeId)
-            };
+                var agentFirstName = InsertNameTextButton.Text;
+                var agentLastName = InsertSurnameTextBox.Text;
+                var agentBirthdate = InsertBirthTextBox.Text;
+                var TypeId = InsertRankTextBox.Text;
 
-            _agents.Create(newAgent);
-            _agents.Save();
+                FBIagentsList newAgent = new FBIagentsList()
+                {
+                    FirstName = agentFirstName,
+                    LastName = agentLastName,
+                    Birthdate = Convert.ToDateTime(agentBirthdate),
+                    AgentTypeId = Int32.Parse(TypeId)
+                };
 
-            LoadAgents();
+                _agents.Create(newAgent);
+                _agents.Save();
+
+                LoadAgents();
+            }
+            else
+            {
+                MessageBox.Show("Insert All Data");
+            }
         }
 
         private void UpdateButton_Click(object sender, EventArgs e)
@@ -76,16 +83,24 @@ namespace ZadDomLab4JakubKazimierski
 
         private void DeleteButton_Click(object sender, EventArgs e)
         {
-            var selectedAgentIndex = Int32.Parse(DeleteIdTextBox.Text);
-            var deleteAgent = _agents.GetById(selectedAgentIndex);
+
+            if (DeleteIdTextBox.Text != "")
+            {
+                var selectedAgentIndex = Int32.Parse(DeleteIdTextBox.Text);
+                var deleteAgent = _agents.GetById(selectedAgentIndex);
 
 
-            _agents.DeleteById(deleteAgent.Id);
-            _agents.Save();
+                _agents.DeleteById(deleteAgent.Id);
+                _agents.Save();
 
 
 
-            LoadAgents();
+                LoadAgents();
+            }
+            else
+            {
+                MessageBox.Show("Insert ID");
+            }
         }
     }
 }
